requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
absl-py==1.2.0; python_version >= '3.6'
anyio==3.6.1; python_full_version >= '3.6.2'
attrs==22.1.0; python_version >= '3.5'
click==8.1.3; python_version >= '3.7'
colorama==0.4.5; platform_system == 'Windows'
cycler==0.11.0; python_version >= '3.6'
fastapi==0.79.1
fonttools==4.36.0; python_version >= '3.7'
h11==0.13.0; python_version >= '3.6'
idna==3.3; python_version >= '3.5'
kiwisolver==1.4.4; python_version >= '3.7'
matplotlib==3.5.3; python_version >= '3.7'
mediapipe==0.8.10.1
numpy==1.23.2; python_version >= '3.7'
opencv-contrib-python==4.6.0.66; python_version >= '3.6'
opencv-python==4.6.0.66
packaging==21.3; python_version >= '3.6'
pillow==9.2.0
protobuf==3.20.1; python_version >= '3.7'
pydantic==1.9.2; python_full_version >= '3.6.1'
pyparsing==3.0.9; python_full_version >= '3.6.8'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.2.0; python_version >= '3.5'
starlette==0.19.1; python_version >= '3.6'
typing-extensions==4.3.0; python_version < '3.10'
uvicorn==0.18.2
